Split addons and plugin tests into separate files
